+++ src/components/POLine/Item/ItemForm.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 import {
@@ -183,7 +183,7 @@ class ItemForm extends Component {
     const instanceId = get(formValues, 'instanceId');
     const isPackage = get(formValues, 'isPackage');
     const title = (
-      <Label required={required}>
+      <Label className={css.titleLabel} marginBottom0 required={required} tagName="div">
         {
           isPackage
             ? <FormattedMessage id="ui-orders.itemDetails.packageName" />
@@ -192,7 +192,7 @@ class ItemForm extends Component {
       </Label>
     );
     const connectedTitle = (
-      <Fragment>
+      <>
         {title}
         <Link
           data-test-connected-link
@@ -205,16 +205,16 @@ class ItemForm extends Component {
             icon="external-link"
           />
         </Link>
-      </Fragment>
+      </>
     );
     const notConnectedTitle = (
-      <Fragment>
+      <>
         {title}
         <div>
           <FormattedMessage id="ui-orders.itemDetails.notConnectedTitle" />
           <InfoPopover content={<FormattedMessage id="ui-orders.itemDetails.notConnectedInfo" />} />
         </div>
-      </Fragment>
+      </>
     );
 
     if (!initialValues.instanceId && !this.state.instanceId) {
@@ -235,7 +235,7 @@ class ItemForm extends Component {
     const isSelectInstanceVisible = !(get(formValues, 'isPackage') || isPostPendingOrder);
 
     return (
-      <Fragment>
+      <>
         <Row>
           <Col
             xs={6}
@@ -257,20 +257,18 @@ class ItemForm extends Component {
             xs={6}
             md={3}
           >
-            <div className={css.titleWrapper}>
-              <Field
-                component={TextField}
-                fullWidth
-                label={this.getTitleLabel()}
-                name="titleOrPackage"
-                onChange={this.setTitleOrPackage}
-                validate={required ? validateRequired : undefined}
-                disabled={isPostPendingOrder}
-              />
-              <div className={css.addButton}>
-                {isSelectInstanceVisible && <InstancePlugin addInstance={this.onAddInstance} />}
-              </div>
-            </div>
+            <Field
+              className={css.titleWrapper}
+              component={TextField}
+              fullWidth
+              label={this.getTitleLabel()}
+              marginBottom0
+              name="titleOrPackage"
+              onChange={this.setTitleOrPackage}
+              validate={required ? validateRequired : undefined}
+              disabled={isPostPendingOrder}
+            />
+            {isSelectInstanceVisible && <InstancePlugin addInstance={this.onAddInstance} />}
           </Col>
           <Col
             xs={6}
@@ -392,7 +390,7 @@ class ItemForm extends Component {
             />
           </Col>
         </Row>
-      </Fragment>
+      </>
     );
   }
 }